+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- search for Form Element Lookups on App open
+
 ## [6.0.0] - 2023-07-26
 
 ### Added